requirements.txt
# Dependencias de Python para el proyecto AWS-Confluent

# Kafka cliente para Confluent Cloud
kafka-python==2.0.2

# Cliente AWS
boto3==1.34.34
botocore==1.34.34

# Variables de entorno
python-dotenv==1.0.0

# Logging y utilidades
colorama==0.4.6

# Desarrollo y testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Formato y calidad de código
black==23.12.1
flake8==6.1.0

# Documentación
markdown==3.5.2

# Serialización de datos
avro-python3==1.11.3

# HTTP requests (para APIs)
requests==2.31.0

# Análisis de datos (opcional)
pandas==2.1.4
numpy==1.26.2

# Monitoreo y métricas
prometheus-client==0.19.0
